文章、大转盘相关代码上传。

+++ Apps/M/Model/MemberPayModel.class.php
@@ -165,7 +165,7 @@ class MemberPayModel extends BaseModel {
 		}
 		else
 		{
-			$payamount=(int)$dataInfo["amount"];
+			$payamount=(float)$dataInfo["amount"];
 			
 			$netpayamount=$payamount;
 			$accountmoney=0;
